1、删除专栏的时候如果有教程，则不删除，教程的 ID 改为统一的 Long 型；3、查询专栏的时候，如果为空则不再查询用户

+++ paicoding-web/src/main/java/com/github/paicoding/forum/web/admin/rest/ColumnSettingRestController.java
@@ -78,14 +78,14 @@ public class ColumnSettingRestController {
 
     @Permission(role = UserRole.ADMIN)
     @GetMapping(path = "deleteColumn")
-    public ResVo<String> deleteColumn(@RequestParam(name = "columnId") Integer columnId) {
+    public ResVo<String> deleteColumn(@RequestParam(name = "columnId") Long columnId) {
         columnSettingService.deleteColumn(columnId);
         return ResVo.ok("ok");
     }
 
     @Permission(role = UserRole.ADMIN)
     @GetMapping(path = "deleteColumnArticle")
-    public ResVo<String> deleteColumnArticle(@RequestParam(name = "id") Integer id) {
+    public ResVo<String> deleteColumnArticle(@RequestParam(name = "id") Long id) {
         columnSettingService.deleteColumnArticle(id);
         return ResVo.ok("ok");
     }
